move APIKeyManager to security module and clean up imports in CRUD operations
